requirements.txt
Flask==2.2.4
requests==2.31.0
deep-translator==1.8.3
python-dotenv==1.0.0
gunicorn==20.1.0
prometheus-client==0.16.0
Werkzeug==2.3.7


